Bluetooth: controller: Refactor LLCP implementation
Separate the LLCP implementation into smaller more manageable parts:

* ull_llcp.c​
  - High level interface​
  - Resource management​
  - Other common code​

* ull_llcp_local.c​
  - Local Request Machine​

* ull_llcp_remote.c​
  - Remote Request Machine​

* ull_llcp_pdu.c​
  - PDU encode/decode​

* ull_llcp_common.c​
  - Common simple procedures (Version Exchange​, Feature Exchange​, etc.)

* ull_llcp_enc.c​
  - Encryption procedure

* ull_llcp_phy.c​
  - PHY Update procedure

Add missing private internal interfaces between the different parts
under the ull_cp_priv namespace.
