+++ tests/bluetooth/controller/common/defaults_cmake.txt
@@ -17,6 +17,14 @@ FILE(GLOB ll_sw_sources
         $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/util/mem.c
         $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/util/memq.c
         $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_tx_queue.c
+
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp_local.c
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp_remote.c
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp_pdu.c
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp_common.c
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp_phy.c
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp_enc.c
+        $ENV{ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/ull_llcp.c
 )
 
 FILE(GLOB mock_sources
@@ -27,3 +35,6 @@ FILE(GLOB common_sources
         $ENV{ZEPHYR_BASE}/tests/bluetooth/controller/common/src/helper_pdu.c
         $ENV{ZEPHYR_BASE}/tests/bluetooth/controller/common/src/helper_util.c
 )
+
+add_definitions(-include kconfig.h)
+add_definitions(-include ztest.h)
