+++ tests/bluetooth/controller/ctrl_api/prj.conf
@@ -1,6 +1,6 @@
-CONFIG_NET_BUF=y
+CONFIG_TEST=y
 CONFIG_ZTEST=y
-CONFIG_ZTEST_ASSERT_VERBOSE=3
+CONFIG_ZTEST_ASSERT_VERBOSE=1
 CONFIG_ZTEST_STACKSIZE=4096
 CONFIG_ZTEST_MOCKING=y
 CONFIG_ZTEST_PARAMETER_COUNT=32
