+++ tests/bluetooth/controller/ctrl_version/src/main.c
@@ -10,292 +10,31 @@
 
 /* Kconfig Cheats */
 
-#define ULL_LLCP_UNITTEST
+#include <bluetooth/hci.h>
+#include <sys/byteorder.h>
+#include <sys/slist.h>
+#include <sys/util.h>
+#include "hal/ccm.h"
 
-/* Implementation Under Test Begin */
-#include "ll_sw/ull_llcp.c"
-/* Implementation Under Test End */
+#include "util/mem.h"
+#include "util/memq.h"
+
+#include "pdu.h"
+#include "ll.h"
+#include "ll_settings.h"
+
+#include "lll.h"
+#include "lll_conn.h"
+
+#include "ull_conn_types.h"
+#include "ull_tx_queue.h"
+#include "ull_llcp.h"
+#include "ull_llcp_internal.h"
 
 #include "helper_pdu.h"
 #include "helper_util.h"
 
 struct ull_cp_conn conn;
-
-/*
- * Note API and internal test are not yet split out here
- */
-
-void test_api_init(void)
-{
-	ull_cp_init();
-	ull_tx_q_init(&conn.tx_q);
-
-	ull_cp_conn_init(&conn);
-	zassert_equal(conn.llcp.local.state, LR_STATE_DISCONNECT, NULL);
-	zassert_equal(conn.llcp.remote.state, RR_STATE_DISCONNECT, NULL);
-}
-
-void test_int_mem_proc_ctx(void)
-{
-	struct proc_ctx *ctx1;
-	struct proc_ctx *ctx2;
-
-	ull_cp_init();
-
-	for (int i = 0U; i < PROC_CTX_BUF_NUM; i++) {
-		ctx1 = proc_ctx_acquire();
-
-		/* The previous acquire should be valid */
-		zassert_not_null(ctx1, NULL);
-	}
-
-	ctx2 = proc_ctx_acquire();
-
-	/* The last acquire should fail */
-	zassert_is_null(ctx2, NULL);
-
-	proc_ctx_release(ctx1);
-	ctx1 = proc_ctx_acquire();
-
-	/* Releasing returns the context to the avilable pool */
-	zassert_not_null(ctx1, NULL);
-}
-
-void test_int_mem_tx(void)
-{
-	bool peek;
-	struct node_tx *tx;
-	struct node_tx *txl[TX_CTRL_BUF_NUM];
-
-	ull_cp_init();
-
-	for (int i = 0U; i < TX_CTRL_BUF_NUM; i++) {
-		peek = tx_alloc_is_available();
-
-		/* The previous tx alloc peek should be valid */
-		zassert_true(peek, NULL);
-
-		txl[i] = tx_alloc();
-
-		/* The previous alloc should be valid */
-		zassert_not_null(txl[i], NULL);
-	}
-
-	peek = tx_alloc_is_available();
-
-	/* The last tx alloc peek should fail */
-	zassert_false(peek, NULL);
-
-	tx = tx_alloc();
-
-	/* The last tx alloc should fail */
-	zassert_is_null(tx, NULL);
-
-	/* Release all */
-	for (int i = 0U; i < TX_CTRL_BUF_NUM; i++) {
-		tx_release(txl[i]);
-	}
-
-	for (int i = 0U; i < TX_CTRL_BUF_NUM; i++) {
-		peek = tx_alloc_is_available();
-
-		/* The previous tx alloc peek should be valid */
-		zassert_true(peek, NULL);
-
-		txl[i] = tx_alloc();
-
-		/* The previous alloc should be valid */
-		zassert_not_null(txl[i], NULL);
-	}
-
-	peek = tx_alloc_is_available();
-
-	/* The last tx alloc peek should fail */
-	zassert_false(peek, NULL);
-
-	tx = tx_alloc();
-
-	/* The last tx alloc should fail */
-	zassert_is_null(tx, NULL);
-
-	/* Release all */
-	for (int i = 0U; i < TX_CTRL_BUF_NUM; i++) {
-		tx_release(txl[i]);
-	}
-}
-
-void test_int_mem_ntf(void)
-{
-	bool peek;
-	struct node_rx_pdu *ntf;
-	struct node_rx_pdu *ntfl[NTF_BUF_NUM];
-
-	ull_cp_init();
-
-	for (int i = 0U; i < NTF_BUF_NUM; i++) {
-		peek = ntf_alloc_is_available();
-
-		/* The previous ntf alloc peek should be valid */
-		zassert_true(peek, NULL);
-
-		ntfl[i] = ntf_alloc();
-
-		/* The previous alloc should be valid */
-		zassert_not_null(ntfl[i], NULL);
-	}
-
-	peek = ntf_alloc_is_available();
-
-	/* The last ntf alloc peek should fail */
-	zassert_false(peek, NULL);
-
-	ntf = ntf_alloc();
-
-	/* The last ntf alloc should fail */
-	zassert_is_null(ntf, NULL);
-
-	/* Release all */
-	for (int i = 0U; i < NTF_BUF_NUM; i++) {
-		ntf_release(ntfl[i]);
-	}
-
-	for (int i = 0U; i < NTF_BUF_NUM; i++) {
-		peek = ntf_alloc_is_available();
-
-		/* The previous ntf alloc peek should be valid */
-		zassert_true(peek, NULL);
-
-		ntfl[i] = ntf_alloc();
-
-		/* The previous alloc should be valid */
-		zassert_not_null(ntfl[i], NULL);
-	}
-
-	peek = ntf_alloc_is_available();
-
-	/* The last ntf alloc peek should fail */
-	zassert_false(peek, NULL);
-
-	ntf = ntf_alloc();
-
-	/* The last ntf alloc should fail */
-	zassert_is_null(ntf, NULL);
-}
-
-void test_int_create_proc(void)
-{
-	struct proc_ctx *ctx;
-
-	ull_cp_init();
-	ull_tx_q_init(&conn.tx_q);
-	ull_cp_conn_init(&conn);
-
-	ctx = create_procedure(PROC_VERSION_EXCHANGE);
-	zassert_not_null(ctx, NULL);
-
-	zassert_equal(ctx->proc, PROC_VERSION_EXCHANGE, NULL);
-	zassert_equal(ctx->state, LP_COMMON_STATE_IDLE, NULL);
-	zassert_equal(ctx->collision, 0, NULL);
-	zassert_equal(ctx->pause, 0, NULL);
-
-	for (int i = 0U; i < PROC_CTX_BUF_NUM; i++) {
-		zassert_not_null(ctx, NULL);
-		ctx = create_procedure(PROC_VERSION_EXCHANGE);
-	}
-
-	zassert_is_null(ctx, NULL);
-}
-
-void test_int_pending_requests(void)
-{
-	struct proc_ctx *peek_ctx;
-	struct proc_ctx *dequeue_ctx;
-	struct proc_ctx ctx;
-
-	ull_cp_init();
-	ull_tx_q_init(&conn.tx_q);
-	ull_cp_conn_init(&conn);
-
-	/* Local */
-
-	peek_ctx = lr_peek(&conn);
-	zassert_is_null(peek_ctx, NULL);
-
-	dequeue_ctx = lr_dequeue(&conn);
-	zassert_is_null(dequeue_ctx, NULL);
-
-	lr_enqueue(&conn, &ctx);
-	peek_ctx = (struct proc_ctx *) sys_slist_peek_head(&conn.llcp.local.pend_proc_list);
-	zassert_equal_ptr(peek_ctx, &ctx, NULL);
-
-	peek_ctx = lr_peek(&conn);
-	zassert_equal_ptr(peek_ctx, &ctx, NULL);
-
-	dequeue_ctx = lr_dequeue(&conn);
-	zassert_equal_ptr(dequeue_ctx, &ctx, NULL);
-
-	peek_ctx = lr_peek(&conn);
-	zassert_is_null(peek_ctx, NULL);
-
-	dequeue_ctx = lr_dequeue(&conn);
-	zassert_is_null(dequeue_ctx, NULL);
-
-	/* Remote */
-
-	peek_ctx = rr_peek(&conn);
-	zassert_is_null(peek_ctx, NULL);
-
-	dequeue_ctx = rr_dequeue(&conn);
-	zassert_is_null(dequeue_ctx, NULL);
-
-	rr_enqueue(&conn, &ctx);
-	peek_ctx = (struct proc_ctx *) sys_slist_peek_head(&conn.llcp.remote.pend_proc_list);
-	zassert_equal_ptr(peek_ctx, &ctx, NULL);
-
-	peek_ctx = rr_peek(&conn);
-	zassert_equal_ptr(peek_ctx, &ctx, NULL);
-
-	dequeue_ctx = rr_dequeue(&conn);
-	zassert_equal_ptr(dequeue_ctx, &ctx, NULL);
-
-	peek_ctx = rr_peek(&conn);
-	zassert_is_null(peek_ctx, NULL);
-
-	dequeue_ctx = rr_dequeue(&conn);
-	zassert_is_null(dequeue_ctx, NULL);
-}
-
-void test_api_connect(void)
-{
-	ull_cp_init();
-	ull_tx_q_init(&conn.tx_q);
-	ull_cp_conn_init(&conn);
-
-	ull_cp_state_set(&conn, ULL_CP_CONNECTED);
-	zassert_equal(conn.llcp.local.state, LR_STATE_IDLE, NULL);
-	zassert_equal(conn.llcp.remote.state, RR_STATE_IDLE, NULL);
-}
-
-void test_api_disconnect(void)
-{
-	ull_cp_init();
-	ull_tx_q_init(&conn.tx_q);
-	ull_cp_conn_init(&conn);
-
-	ull_cp_state_set(&conn, ULL_CP_DISCONNECTED);
-	zassert_equal(conn.llcp.local.state, LR_STATE_DISCONNECT, NULL);
-	zassert_equal(conn.llcp.remote.state, RR_STATE_DISCONNECT, NULL);
-
-	ull_cp_state_set(&conn, ULL_CP_CONNECTED);
-	zassert_equal(conn.llcp.local.state, LR_STATE_IDLE, NULL);
-	zassert_equal(conn.llcp.remote.state, RR_STATE_IDLE, NULL);
-
-	ull_cp_state_set(&conn, ULL_CP_DISCONNECTED);
-	zassert_equal(conn.llcp.local.state, LR_STATE_DISCONNECT, NULL);
-	zassert_equal(conn.llcp.remote.state, RR_STATE_DISCONNECT, NULL);
-}
-
-
 
 static void setup(void)
 {
@@ -511,20 +250,6 @@ void test_version_exchange_mas_rem_2(void)
 
 void test_main(void)
 {
-	ztest_test_suite(internal,
-			 ztest_unit_test(test_int_mem_proc_ctx),
-			 ztest_unit_test(test_int_mem_tx),
-			 ztest_unit_test(test_int_mem_ntf),
-			 ztest_unit_test(test_int_create_proc),
-			 ztest_unit_test(test_int_pending_requests)
-			);
-
-	ztest_test_suite(public,
-			 ztest_unit_test(test_api_init),
-			 ztest_unit_test(test_api_connect),
-			 ztest_unit_test(test_api_disconnect)
-			);
-
 	ztest_test_suite(version_exchange,
 			 ztest_unit_test_setup_teardown(test_version_exchange_mas_loc, setup, unit_test_noop),
 			 ztest_unit_test_setup_teardown(test_version_exchange_mas_loc_2, setup, unit_test_noop),
@@ -532,7 +257,5 @@ void test_main(void)
 			 ztest_unit_test_setup_teardown(test_version_exchange_mas_rem_2, setup, unit_test_noop)
 			);
 
-	ztest_run_test_suite(internal);
-	ztest_run_test_suite(public);
 	ztest_run_test_suite(version_exchange);
 }
