--- tests/bluetooth/ctrl_llcp/testcase.yaml
@@ -1,5 +1,0 @@
-common:
-  tags: bluetooth
-tests:
-  bluetooth.ctrl_llcp.test:
-    platform_whitelist: native_posix
